Add bumpers/bounding to canvas
Large commit that not only adds collision detection for obstacles but also fixes
a few bugs and refactors how bumpers are constructed.

Collision detection has two main loops (which eventually can be refactored out):

* The first loop looks for overlap between a given obstacle and the ball in
both X and Y.
* If a collision is detected, the next step is determining which direction the
ball came from by calculating the previous position by reversing the
speed/direction math.
* These points are then used to ask if the ball is coming from the top or
the bottom, and if from the left or the right
* Theoretically a ball could have a collision from both the X and Y direction
simultaneously (and thus each check is an `if` not an `else if`) by hitting
the corner perfectly
* When either direction is detected, its cooresponding direction is reversed

Because this detection relies on X max or min and Y max or min, the Bumper
constructor was altered to take an X max and Y max instead of a width/height.
This also allows the resulting list of arguments to be passed directly into
the obstacle array, though eventually this should be generated automatically.

The last refactor concerned how objects were being drawn. The animationFrame
now draws all the obstacles, and the fillStyle is set before it's drawn.