Add level creation, gameLoop includes level
Added method to start a level from the game engine portion of the index.

While currently a level doesn't have an end condition and the engine will need
to incorporate a method to update which level is being played, it should be
generally easy to customize and create levels now.

A new level is defined in the levels directory. Bumpers are placed in their own
directory, since bumper definition is a bit bulky, but the ball, par, and hole
information is defined in a single file for all levels. Just add a new key that
represents the number of the level/hole.

Closes #36