Ball now ALWYAS goes the way you expect
* I tracked down this problem, and in a weird moment of realization figured out the simple fix.

The ball was doing a weird thing where every odd numbered bounce off of a wall it would bounce in the wrong direction. This made me think it had something to do with the check where we multiply x/yDirection by -1.
I couldn't figure out how to only swap the number every other time though, and it didn't really make sense.
Then I realized exactly what we were doing.
By leaving the direction negative, when you clicked the next time, the slope would calculate correctly, but the movement is based on a positive speed, initially. This made the ball seem to move backwards.
I added a step in the stopCheck function to reset the directions to positive 1. This makes everything behave as expected.