Add sand obstacles
Added sand as a type of bumper. This necessitated changing the collision
detection methods to distinguish which type of obstacle was being called. Please
note that now you will have to specify that a bumper is of type "bumper" in the
hash format to create bumpers, either in tests or in the game. This should
facilitate adding in hills as well.